solve bug

+++ data_sample_utils.py
@@ -10,8 +10,8 @@ def random_split(data_list, ratio=None, num=None):
         if num > len(data_list):
             assert False
     elif ratio != None:
-        num = ratio * len(data_list)
-
+        num = int(ratio * len(data_list))
+        
     idx_list = [i for i, _ in enumerate(data_list)]
     random.shuffle(idx_list)
     sample_list = [data_list[i] for i in idx_list[:num]]
